separate input and button into their own component
